+++ tests/Main.hs
@@ -170,6 +170,8 @@ encodeUnitTests = testGroup "Encoder unit tests"
   , encodeBytesFromBuilder
   , encodeMessageReflection
   , encodeCachedSubmessage
+  , testShowMessageEncoder
+  , testShowMessageEncoding
   ]
 
 -- TODO: We should consider generating the reference encodings
@@ -334,17 +336,13 @@ encodeCachedSubmessage = testGroup "Cached Submessages"
   [ testOptional
   , testRepeated
   , testOneof
-  , testShow
   ]
   where
-    cacheReflection :: forall a . Message a => a -> FormE.MessageEncoding a
-    cacheReflection = FormE.cacheMessageEncoding . FormE.messageReflection
-
     testOptional :: TestTree
     testOptional = testCase "cached optional submessage" $ do
       let trivial = TP.Trivial 123
           wrappedTrivial = FormE.fieldsToMessage @TP.WrappedTrivial $
-            FormE.field @"trivial" (Just (cacheReflection trivial))
+            FormE.field @"trivial" (Just (FormE.messageCache trivial))
       FormE.toLazyByteString wrappedTrivial @?=
         toLazyByteString (TP.WrappedTrivial (Just trivial))
 
@@ -356,7 +354,7 @@ encodeCachedSubmessage = testGroup "Cached Submessages"
             , TP.MapTestEmulation_Trivial 25 (Just (TP.WrappedTrivial (Just (TP.Trivial 789))))
             ]
           mapTestEmulation = FormE.fieldsToMessage @TP.MapTestEmulation $
-            FormE.field @"trivial" (V.map cacheReflection trivials)
+            FormE.field @"trivial" (V.map FormE.messageCache trivials)
       FormE.toLazyByteString mapTestEmulation @?=
         toLazyByteString (TP.MapTestEmulation mempty trivials mempty)
 
@@ -364,29 +362,39 @@ encodeCachedSubmessage = testGroup "Cached Submessages"
     testOneof = testCase "cached submessage within oneof" $ do
       let withOneof = TPOI.WithOneof (Just (TPOI.WithOneofPickOneB 123))
           withImported = FormE.fieldsToMessage @TPO.WithImported $
-            FormE.field @"withOneof" (cacheReflection withOneof)
+            FormE.field @"withOneof" (FormE.messageCache withOneof)
       FormE.toLazyByteString withImported @?=
         toLazyByteString (TPO.WithImported (Just (TPO.WithImportedPickOneWithOneof withOneof)))
 
-    testShow :: TestTree
-    testShow = testCase "Show MessageEncoding" $ do
-      let trivial = TP.Trivial 123
-          badForm = "abc"
-          encoding :: FormE.MessageEncoding TP.Trivial
-          encoding = cacheReflection trivial
-          bogus :: FormE.MessageEncoding TP.Trivial
-          bogus = FormE.unsafeByteStringToMessageEncoding badForm
-      -- Test desired behavior for valid encoding:
-      show encoding @?= "Proto3.Suite.Form.Encode.cacheMessage " ++ showsPrec 11 trivial ""
-      -- Check that when interpreting the above expected 'show' output
-      -- as Haskell source, it really does produce the shown value:
-      encoding @=? FormE.cacheMessage trivial
-      -- Test fallback behavior for invalid encoding:
-      show bogus @?=
-        "Proto3.Suite.Form.Encode.unsafeByteStringToMessageEncoding " ++ showsPrec 11 badForm ""
-      -- Check that when interpreting the above expected 'show' output
-      -- as Haskell source, it really does produce the shown value:
-      bogus @=? FormE.unsafeByteStringToMessageEncoding badForm
+testShowMessageEncoding :: TestTree
+testShowMessageEncoding = testCase "Show MessageEncoding" $ do
+  let trivial = TP.Trivial 123
+      badForm = "abc"
+      encoding :: FormE.MessageEncoding TP.Trivial
+      encoding = FormE.messageCache trivial
+      bogus :: FormE.MessageEncoding TP.Trivial
+      bogus = FormE.unsafeByteStringToMessageEncoding badForm
+  -- Test desired behavior for valid encoding:
+  show (Just encoding) @?=
+    "Just (Proto3.Suite.Form.Encode.messageCache " ++ showsPrec 11 trivial ")"
+  -- Test fallback behavior for invalid encoding:
+  show (Just bogus) @?=
+    "Just (Proto3.Suite.Form.Encode.unsafeByteStringToMessageEncoding " ++ showsPrec 11 badForm ")"
+
+testShowMessageEncoder :: TestTree
+testShowMessageEncoder = testCase "Show MessageEncoder" $ do
+  let trivial = TP.Trivial 123
+      badForm = "abc"
+      encoder :: FormE.MessageEncoder TP.Trivial
+      encoder = FormE.messageReflection trivial
+      bogus :: FormE.MessageEncoder TP.Trivial
+      bogus = FormE.unsafeByteStringToMessageEncoder badForm
+  -- Test desired behavior for valid encoder:
+  show (Just encoder) @?=
+    "Just (Proto3.Suite.Form.Encode.messageReflection " ++ showsPrec 11 trivial ")"
+  -- Test fallback behavior for invalid encoder:
+  show (Just bogus) @?=
+    "Just (Proto3.Suite.Form.Encode.unsafeByteStringToMessageEncoder " ++ showsPrec 11 badForm ")"
 
 data TestMessage
        (num :: Nat)